Edited MASK section of docs
Appreciate the effort put into the documentation and site. It 
deserves more visibility. I often see people on Reddit/Discord looking for good docs for creating nodes.

 I've suggested some edits for the MASK section.

There was some things that I thought might be slightly incorrect, but I could be wrong. The documentation states that masks have a default shape of `[B, H, W, C]`, but my understanding is that masks are created by calling `numpy.array()` on the alpha channel of a `PIL.Image.Image`, which creates a 2D array since it's a single channel. When it's then converted to a tensor, it will be `[B, H, W]`. 

Although I do remember at one point being in a scenario where I was getting `[B, H, W, C]` masks, so I understand why it was written. 

There's also note at the bottom of the page that also says MASKs are channel-first, but I don't believe MASKs have a channel dim by default and if we are to expand we should expand channel-last to match the expected type of IMAGE. 

In my suggested changes I added some snippets that can help people do common operations with masks. 

Thanks so much for making these docs! Everyone whose contributed did a great job explaining all the importatn stuff  concisely. Looking forward to your feedback on these changes.