Add security standard: No use of `requests` to download Python code/libraries for use in modules
This is in the same spirit as the restriction of not allowing `pip` calls via `subprocess` calls.

The reason this is being suggested now is due to this observation of a module in the ComfyUI registry doing just this - using `requests` to download Python code from a separate repository and then write those files and then in the same module usage, import that library.  This is dangerous because it is unvetted / unauditable code.  https://github.com/1038lab/ComfyUI-OmniGen/blob/4f9d6a945e3fa7aaf4485c4e7c5292c9b4826ab7/ailab_OmniGen.py#L76